Website Redo
Redid the website. Improving the css and html markup. New experimental design. Included the SWUP framework to make page transitions more seamless. Used the GSAP library to v3.x doc guidelines. Added separate pages for each of the four projects. Removed the last project that linked to my behance (it was useless). Created project pages that show off the project images alongside an overview and tools used with a link to the project/website/github. And other such stuff.