Update project order and about me section
Changed the order of the projects. Moved the constructed images project to the bottom to show off more of my developmental projects. Employers more likely to view first project as the better. Even though all are great, the others relate more to development.

Edited the about me section to conform with my new resume more than it used to. Showing skills that I know.